requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
google-genai==0.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
